contributions/antlrTrees/src/test/java/org/softlang/company/tests/TotalTest.java
package org.softlang.company.tests;

import static org.softlang.company.features.Parsing.*;
import static org.softlang.company.features.Total.*;

import org.antlr.runtime.*;
import org.antlr.runtime.tree.CommonTree;
import org.antlr.runtime.tree.CommonTreeNodeStream;

import java.io.File;
import java.io.IOException;

import static org.junit.Assert.assertEquals;
import org.junit.Test;

public class TotalTest {

    private static String sampleCompany =
        "inputs"
        + File.separator
        + "sample.Company";

    @Test
    public void testTotal()
            throws RecognitionException, IOException {
        CommonTree tree = parse(sampleCompany);
        double total = total(tree);
        assertEquals(399747, total, 0);
    }

}